move create optimizer and scheduler function to avoid name conflict with utils
