+++ platalea/asr.py
@@ -5,7 +5,6 @@ import numpy as np
 import torch
 import torch.nn as nn
 import torch.nn.functional as F
-import torch.optim as optim
 
 import platalea.schedulers
 import platalea.dataset as D
@@ -14,7 +13,8 @@ from platalea.encoders import SpeechEncoder
 import platalea.loss
 import platalea.score
 import platalea.hardware
-from platalea.utils import create_optimizer, create_scheduler
+from platalea.optimizers import create_optimizer
+from platalea.schedulers import create_scheduler
 
 
 class SpeechTranscriber(nn.Module):
