+++ platalea/mtl.py
@@ -3,7 +3,6 @@ import json
 import logging
 import torch
 import torch.nn as nn
-import torch.optim as optim
 
 import platalea.schedulers
 from platalea.encoders import SpeechEncoderBottom, SpeechEncoderSplit
@@ -13,7 +12,8 @@ from platalea.asr import SpeechTranscriber
 import platalea.loss
 import platalea.score
 import platalea.hardware
-from platalea.utils import create_optimizer, create_scheduler
+from platalea.optimizers import create_optimizer
+from platalea.schedulers import create_scheduler
 
 
 class MTLNetASR(nn.Module):
