+++ platalea/schedulers.py
@@ -49,3 +49,22 @@ def constant(optimizer, lr):
         return lr
 
     return lr_scheduler.LambdaLR(optimizer, learning_rate, last_epoch=-1)
+
+
+def create_scheduler(config, optimizer, data):
+    if 'lr' in config.keys():
+        raise KeyError('Illegal keyword "lr" used in config. Use keyword "constant_lr" instead.')
+
+    configured_scheduler = config.get('lr_scheduler')
+
+    if configured_scheduler is None or configured_scheduler == 'cyclic':
+        scheduler = platalea.schedulers.cyclic(optimizer, len(data['train']), max_lr=config['max_lr'],
+                                               min_lr=config['min_lr'])
+    elif configured_scheduler == 'noam':
+        scheduler = platalea.schedulers.noam(optimizer, config['d_model'])
+    elif configured_scheduler == 'constant':
+        scheduler = platalea.schedulers.constant(optimizer, config['constant_lr'])
+    else:
+        raise Exception(
+            "lr_scheduler config value " + configured_scheduler + " is invalid, use cyclic or noam or constant")
+    return scheduler
